refactor: move remaining permission APIs from shared into sdk
