+++ libs/shared/package.json
@@ -6,7 +6,6 @@
     "./globals.css": "./src/styles/globals.css",
     "./flexlayout.css": "./src/components/flex-layout/flexlayout.css",
     ".": "./src/index.ts",
-    "./api": "./src/api/index.ts",
     "./assets/img/*": "./src/assets/img/*"
   },
   "license": "MIT",
