--- libs/shared/src/api/index.ts
@@ -1,1 +1,0 @@
-export * from "./permissions";
