+++ libs/shared/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./api";
 export * from "./components";
 export * from "./constants";
 export * from "./hooks";
