+++ packages/client/src/pages/share-page.tsx
@@ -1,8 +1,8 @@
 import { observer } from "mobx-react-lite";
 import { lazy, Suspense, useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
+import { getUserProjectPermission as getUserProjectLevelPermission } from "@semoss/sdk";
 import { runPixel } from "@semoss/sdk/react";
-import { getUserProjectPermission as getUserProjectLevelPermission } from "@semoss/shared";
 import { Spinner, toast } from "@semoss/ui/next";
 import type { AppMetadata, AppType } from "@/components/app";
 import { PlatformMessages } from "@/components/shared";
