+++ packages/client/src/stores/config/config.store.ts
@@ -1,7 +1,6 @@
 import { makeAutoObservable, runInAction } from "mobx";
-// TODO: Pull from sdk
+import { getUserProjectPermission as getUserProjectLevelPermission } from "@semoss/sdk";
 import { Env, logout, runPixel } from "@semoss/sdk/react";
-import { getUserProjectPermission as getUserProjectLevelPermission } from "@semoss/shared/api";
 import { registerUser } from "@/api";
 import type { AppMetadata } from "@/components/app";
 import { THEME } from "@/constants";
