+++ packages/playground/src/components/workspace/members/workspace-members-list.tsx
@@ -5,10 +5,9 @@ import {
 	editProjectUserPermissions,
 	getProjectUsers,
 	getUserProjectPermission,
-	type Role,
 	removeProjectUserPermissions,
-	type User,
-} from "@semoss/shared";
+} from "@semoss/sdk";
+import type { Role, User } from "@semoss/shared";
 import {
 	Button,
 	Dialog,
