+++ packages/playground/src/components/workspace/members/workspace-sharing-modal.tsx
@@ -3,14 +3,10 @@ import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "@semoss/i18n";
 import {
 	addProjectUserPermissions,
+	getProjectUsers,
 	getProjectUsersNoCredentials,
 } from "@semoss/sdk";
-import {
-	getProjectUsers,
-	type PostUser,
-	type Role,
-	type User,
-} from "@semoss/shared";
+import type { PostUser, Role, User } from "@semoss/shared";
 import {
 	Button,
 	Command,
